// Qanat.Web/src/app/shared/generated/model/geography-public-dto.ts
export class GeographyPublicDto { 
    GeographyID?: number;
    GeographyName?: string | null;
    GeographyDisplayName?: string | null;
    GeographyDescription?: string | null;
    StartYear?: number;
    IsDemoGeography?: boolean;
    Color?: string | null;
    AllocationPlansEnabled?: boolean;
    AllocationPlansVisibleToLandowners?: boolean;
    AllocationPlansVisibleToPublic?: boolean;
    ContactEmail?: string | null;
    ContactPhoneNumber?: string | null;
    LandingPageEnabled?: boolean;
    MeterDataEnabled?: boolean;
    WellRegistryEnabled?: boolean;
    FeeCalculatorEnabled?: boolean;
    AllowWaterMeasurementSelfReporting?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}