// Qanat.Web/src/app/shared/generated/model/gdb-columns-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FeatureClassInfo } from './feature-class-info';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class GDBColumnsDto { 
    FileResourceID?: number;
    FeatureClasses?: Array<FeatureClassInfo> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface GDBColumnsDtoForm { 
    FileResourceID?: FormControl<number>;
    FeatureClasses?: FormControl<Array<FeatureClassInfo>>;
}

export class GDBColumnsDtoFormControls { 
    public static FileResourceID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FeatureClasses = (value: FormControlState<Array<FeatureClassInfo>> | Array<FeatureClassInfo> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<FeatureClassInfo>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}