// Qanat.Web/src/app/shared/generated/model/column-mapping-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ColumnMappingDto { 
    SourceColumnName?: string | null;
    DestinationColumnName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ColumnMappingDtoForm { 
    SourceColumnName?: FormControl<string>;
    DestinationColumnName?: FormControl<string>;
}

export class ColumnMappingDtoFormControls { 
    public static SourceColumnName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static DestinationColumnName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}